moviesweb/requirements.txt
Django==4.2.23
djangorestframework==3.15.2
django-cors-headers==4.4.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
Pillow==10.4.0
django-storages==1.14.4
boto3==1.34.162
python-dotenv==1.0.1
